Made changes to the code that correlate with the usage of the PWM signals
Added Timer initialization to the USER CODE 2 BEGIN part for channels 1 and 2 on Timer 3. There is a full description of the project settings that I used for the PWM signal outputs, found within the same block. Adjusted the code within the change_motor_state code to be able to move the bot in the specified direction given the state. May need to further adjust case 3 and case 4.